Add files via upload
Using Js to access Dom Objects
We are calling textChange() function to find text input by the id.  Then we are getting all the elements by tag and class. Looping through the outElements and then we are setting the innerHTML text to InElement.value from what we entered in the text box.